% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mapCircles.R
\name{mapCircles}
\alias{mapCircles}
\title{Proportional Circles}
\usage{
mapCircles(shpFile, shpId, df, dfId, var, fixedNorm = FALSE,
  shareOfCircles = 0.02, radiusMax = 0.5, valueMax = max(df[, var], na.rm
  = TRUE), lgdRnd = 0, posLeg = "bottomleft", circleCol = "#FD8D3C",
  baseCol = "#FFEDA0", title = var, legend = var, author = "author",
  scalebar = FALSE, scalebarSize, scalebarText, northArrow = FALSE,
  northArrowSize, txtCex = 1)
}
\arguments{
\item{shpFile}{Path to a polygon shapefile.}

\item{shpId}{Unique identifier of the shapefile. It must correspond to dfId.}

\item{df}{Data frame that contains the variable to be mapped.}

\item{dfId}{Unique identifier of the data frame. It must correspond to shpId.}

\item{var}{Name of the positive numeric variable to be mapped.}

\item{fixedNorm}{FALSE (default) : the sum of the surfaces occupied by circles is proportional to the size of the map and is declared by the shareOfCircles argument, radiusMax and valueMax are not used.
TRUE : the size of the largest circle is defined by a radius size (radiusMax) and a fixed variable value (valueMax), the shareOfCircles argument is not used.}

\item{shareOfCircles}{Share of the surface of the map occupied by circles (0.02 is 20 percent).}

\item{radiusMax}{Size (in cm) of the radius of the biggest circle.}

\item{valueMax}{Value used to normalize the size of the largest circle (in variable units).}

\item{lgdRnd}{Rounding of the values of the variable presented in the legend.}

\item{posLeg}{Position of the legend (top, bottom, left, right, center, topleft, topright, bottomleft or bottomright).}

\item{circleCol}{Color of the circles.}

\item{baseCol}{Color of the base map.}

\item{title}{Title of the map.}

\item{legend}{Title of the legend.}

\item{author}{Line (author and sources) at the bottom of the map are available to display additional information. It is recommended to display the name of the author and the sources of the data and the base map.}

\item{scalebar}{FALSE (default): don't draw a scale bar.
TRUE : draws a scale bar. The choice of the scale bar location is interactive.}

\item{scalebarSize}{Size of the scale bar in map units.}

\item{scalebarText}{Text of the scale bar.}

\item{northArrow}{FALSE (default) : don't draw a North arrow.
TRUE : draws a North arrow. The choice of the North arrow location is interactive.}

\item{northArrowSize}{Size of the North arrow in map units.}

\item{txtCex}{Size of the texts.}
}
\value{
A proportionnal circle map is displayed in the graphic window.
You can export the map in raster or vector format (pdf).
}
\description{
Proportional Circles
}
\details{
The circles are centered on the centroids of the polygons.
Only the five first arguments are compulsory.
}
\examples{
# library(rCarto)
data(mtq)
# minimal example
mapCircles(shpFile=file.path(path.package("rCarto"), "shapes/COMMUNE"),
           shpId="INSEE_COM",df=mtq,dfId="ID",var="P09_POP", )
}

